requirements.txt
kafka-python
faker
redis
pyspark
numpy